--- app/src/main/java/com/ysshin/fine_dust_app/data/WeatherContract.kt
@@ -1,12 +1,0 @@
-package com.ysshin.fine_dust_app.data
-
-import android.provider.BaseColumns
-
-object WeatherContract {
-    object WeatherEntry: BaseColumns {
-        const val TABLE_NAME = "weather"
-
-    }
-
-
-}
