+++ app/src/main/java/com/ysshin/fine_dust_app/di/PreferenceModule.kt
@@ -1,6 +1,6 @@
 package com.ysshin.fine_dust_app.di
 
-import com.ysshin.fine_dust_app.data.PreferenceManager
+import com.ysshin.fine_dust_app.utils.PreferenceManager
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
