+++ app/src/main/java/com/ysshin/fine_dust_app/ui/RegistrationFragment.kt
@@ -10,7 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.ysshin.fine_dust_app.R
 import com.ysshin.fine_dust_app.data.Auth
-import com.ysshin.fine_dust_app.data.PreferenceManager
+import com.ysshin.fine_dust_app.utils.PreferenceManager
 import com.ysshin.fine_dust_app.databinding.FragmentRegistrationBinding
 import com.ysshin.fine_dust_app.viewmodels.RegistrationViewModel
 import org.koin.android.ext.android.inject
